Add an option to make rustc_wrapper.py dump the command it runs
This is useful for debugging.

Bug: 1500558
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5013308
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1221687}

+++ build/rust/rustc_wrapper.py
@@ -8,6 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 import argparse
 import pathlib
 import subprocess
+import shlex
 import os
 import sys
 import re
@@ -142,6 +143,7 @@ def main():
   parser.add_argument('--depfile', required=True, type=pathlib.Path)
   parser.add_argument('--rsp', type=pathlib.Path, required=True)
   parser.add_argument('--target-windows', action='store_true')
+  parser.add_argument('-v', action='store_true')
   parser.add_argument('args', metavar='ARG', nargs='+')
 
   args = parser.parse_args()
@@ -192,9 +194,13 @@ def main():
     fixed_env_vars.append(k)
 
   try:
+    if args.v:
+      print(' '.join(f'{k}={shlex.quote(v)}' for k, v in env.items()),
+            args.rustc, shlex.join(rustc_args))
     r = subprocess.run([args.rustc, *rustc_args], env=env, check=False)
   finally:
-    os.remove(out_rsp)
+    if not args.v:
+      os.remove(out_rsp)
   if r.returncode != 0:
     sys.exit(r.returncode)
